Getting rid of duplicated code in ITestQueryOperation
