+++ src/test/java/com/basho/riak/client/core/operations/itest/ts/ITestQueryOperation.java
@@ -5,7 +5,6 @@ import com.basho.riak.client.core.operations.ts.QueryOperation;
 import com.basho.riak.client.core.operations.ts.StoreOperation;
 import com.basho.riak.client.core.query.timeseries.QueryResult;
 import com.basho.riak.client.core.util.BinaryValue;
-import junit.framework.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -16,6 +15,7 @@ import static org.junit.Assert.*;
 /**
  * Time Series Query Operation Integration Tests
  * @author Alex Moore <amoore at basho dot com>
+ * @author Sergey Galkin <srggal at gmail dot com>
  * @since 2.0.3
  */
 
@@ -42,13 +42,9 @@ public class ITestQueryOperation extends ITestTsBase
                                  "  and user ='user1'" +
                                  "  and geohash ='hash1'";
 
-        final BinaryValue queryTextBS = BinaryValue.create(queryText);
-
-        QueryOperation queryOp = new QueryOperation.Builder(queryTextBS).build();
-        RiakFuture<QueryResult, BinaryValue> future = cluster.execute(queryOp);
-
-        QueryResult queryResult = future.get();
-        assertTrue(future.isSuccess());
+        final QueryResult queryResult = executeQuery(
+                new QueryOperation.Builder(BinaryValue.create(queryText))
+            );
 
         assertNotNull(queryResult);
         assertEquals(0, queryResult.getColumnDescriptions().size());
@@ -63,13 +59,10 @@ public class ITestQueryOperation extends ITestTsBase
                                  "  and time < "+ now +
                                  "  and user ='user2'" +
                                  "  and geohash ='hash1'";
-        final BinaryValue queryTextBS = BinaryValue.create(queryText);
 
-        QueryOperation queryOp = new QueryOperation.Builder(queryTextBS).build();
-        RiakFuture<QueryResult, BinaryValue> future = cluster.execute(queryOp);
-
-        QueryResult queryResult = future.get();
-        assertTrue(future.isSuccess());
+        final QueryResult queryResult = executeQuery(
+                new QueryOperation.Builder(BinaryValue.create(queryText))
+        );
 
         assertNotNull(queryResult);
         assertEquals(7, queryResult.getColumnDescriptions().size());
